2.8

+++ src/main/java/fr/formiko/formiko/Individu.java
@@ -4,7 +4,7 @@ import fr.formiko.formiko.Main;
 import fr.formiko.usuel.debug;
 import fr.formiko.usuel.erreur;
 import fr.formiko.usuel.g;
-import fr.formiko.usuel.listes.GString;
+import fr.formiko.usuel.structures.listes.GString;
 import fr.formiko.usuel.tableau;
 
 import java.io.Serializable;
@@ -16,7 +16,7 @@ public class Individu implements Serializable{
   protected boolean sexe; // false = femmelle; true = Male;
   protected byte nétoyage;
   protected byte actionMax; // dans certaine espèces les larves travaille en créant du fil de soie.
-  protected byte taille;
+  protected int taille;
   protected String couleur;
   protected byte poidMax; //(en gramme) si le poid est supérieur a 10% la fourmi peine et ses déplacements coute plus chere.
   protected byte tempsDeRepos; // en heure (ou en 24a de tour).
@@ -25,8 +25,8 @@ public class Individu implements Serializable{
   protected int tAgeMax[] = new int[4]; // sauf l'age adulte qui vari.
   protected int tNourritureMax[]= new int[4]; // sauf la nourritureMax adulte.
   protected byte tCoutAction[] = new byte[6];
-  // CONSTRUCTEUR ---------------------------------------------------------------
-  public Individu(int idEsp, byte ty, boolean se,boolean se2, byte né, byte ac, byte ta, String co, byte po, byte te, int tag0, int tag1, int tag2, int tag3, int nm0, int nm1, int nm2, int nm3,byte ca0,byte ca1,byte ca2,byte ca3,byte ca4,byte ca5,int nc){
+  // CONSTRUCTORS --------------------------------------------------------------
+  public Individu(int idEsp, byte ty, boolean se,boolean se2, byte né, byte ac, int ta, String co, byte po, byte te, int tag0, int tag1, int tag2, int tag3, int nm0, int nm1, int nm2, int nm3,byte ca0,byte ca1,byte ca2,byte ca3,byte ca4,byte ca5,int nc){
     e = Main.getGEspece().getEspeceParId(idEsp);
     type = ty; sexué = se; sexe = se2; nétoyage = né; actionMax = ac; taille =ta; couleur = co; poidMax =po; tempsDeRepos = te;
     tAgeMax[0] = tag0; tAgeMax[1] = tag1; tAgeMax[2] = tag2; tAgeMax[3] = tag3;
@@ -35,11 +35,11 @@ public class Individu implements Serializable{
     nourritureConso=nc;
     debug.débogage("Fin de la création d'un individu");
   }
-  // GET SET --------------------------------------------------------------------
+  // GET SET -------------------------------------------------------------------
   public Espece getEspece(){ return e;}
   public byte getType(){ return type;}
   public byte getActionMax(){ return actionMax;}
-  public byte getTaille(){ return taille;}
+  public int getTaille(){ return taille;}
 
   public int getAgeMax(int i){ return tAgeMax[i]/2;} //to make it shorter
   public int getAgeMax(){ return getAgeMax(3);}
@@ -67,7 +67,7 @@ public class Individu implements Serializable{
     if(stade==-3){ return 0;}
     return nourritureConso;
   }
-  // Fonctions propre -----------------------------------------------------------
+  // FUNCTIONS -----------------------------------------------------------------
   public String toString(){
     GString adj = new GString();//♂ ♀
     if(sexué){ String sex = "♀"; if(sexe){ sex = "♂";} adj.add(g.get("sexué")+" ("+sex+")");}
